split out browserStorage utilities
