+++ frontend/package.json
@@ -7,6 +7,7 @@
     "@sentry/react": "^5.27.4",
     "@sentry/tracing": "^5.27.4",
     "@types/chroma-js": "^2.1.2",
+    "@types/lodash": "^4.14.165",
     "@types/node": "^14.14.10",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
